Mengubah path upload foto ke public/upload dan Implementasi Update Produk admin

+++ database/migrations/2025_11_06_021127_create_produks_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
 
             //Produk
             $table->string('nama_produk', '100');
-            $table->float('harga_produk');
+            $table->decimal('harga_produk', 10, 2);
             $table->text('deskripsi_produk');
 
             //Detail Produk
